REVIEW: Remove a useless "drawRect" + fixing a "drawRect" call
Those changes don't have an impact on the final result.
The loop draw the entire image so the drawCall removed was useless.
And drawing a rectangle with a size of 0x0 has no sense.
The library handles that and assume a rect of 1x1 but from the point of view of the reader a rect of 0x0 doesn't make sense.